requirements.txt
Django==1.8.2
Pillow==2.9.0
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
defusedxml==0.4.1
django-allauth==0.21.0
django-bootstrap-form==3.2
django-bootstrap-toolkit==2.15.0
django-celery==3.1.16
kombu==3.0.26
oauthlib==0.7.2

python3-openid==3.0.5
pytz==2015.4
requests-oauthlib==0.5.0
